enimi/requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1.4
django-bootstrap5==22.1
django-environ==0.9.0
django-phonenumber-field==7.0.0
djangorestframework==3.14.0
phonenumbers==8.13.0
Pillow==9.3.0
psycopg2-binary==2.9.5
pytz==2022.7
sqlparse==0.4.3
